refactor: split Utils.kt by function or receiver type
