+++ app/src/main/java/com/osfans/trime/util/AppContext.kt
@@ -1,0 +1,11 @@
+/*
+ * SPDX-FileCopyrightText: 2015 - 2024 Rime community
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+package com.osfans.trime.util
+
+import android.content.Context
+import com.osfans.trime.TrimeApplication
+
+val appContext: Context get() = TrimeApplication.getInstance().applicationContext
