+++ app/src/main/java/com/osfans/trime/util/Bundle.kt
@@ -1,0 +1,18 @@
+/*
+ * SPDX-FileCopyrightText: 2015 - 2024 Rime community
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+package com.osfans.trime.util
+
+import android.os.Build
+import android.os.Bundle
+import java.io.Serializable
+
+inline fun <reified T : Serializable> Bundle.serializable(key: String): T? =
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+        getSerializable(key, T::class.java)
+    } else {
+        @Suppress("DEPRECATION")
+        getSerializable(key) as? T
+    }
