+++ app/src/main/java/com/osfans/trime/util/RecyclerView.kt
@@ -1,0 +1,20 @@
+/*
+ * SPDX-FileCopyrightText: 2015 - 2024 Rime community
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+package com.osfans.trime.util
+
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import androidx.recyclerview.widget.RecyclerView
+
+fun RecyclerView.applyNavBarInsetsBottomPadding() {
+    clipToPadding = false
+    ViewCompat.setOnApplyWindowInsetsListener(this) { _, windowInsets ->
+        windowInsets.getInsets(WindowInsetsCompat.Type.navigationBars()).also {
+            setPadding(paddingLeft, paddingTop, paddingRight, it.bottom)
+        }
+        windowInsets
+    }
+}
